requirements.txt
Django==1.8.3
Fabric==1.10.2
GitPython==1.0.1
PyYAML==3.11
coverage==3.7.1
ecdsa==0.13
gitdb==0.6.4
paramiko==1.15.2
pycrypto==2.6.1
python-coveralls==2.5.0
requests==2.7.0
sh==1.11
six==1.9.0
smmap==0.9.0
wsgiref==0.1.2
